bug fixing in test_strategy.py
